+++ lib/util/cache/package/index.ts
@@ -1,8 +1,8 @@
 import type { GlobalConfig } from '../../../config/types';
 import * as memCache from '../memory';
-import type { PackageCache } from './common';
 import * as fileCache from './file';
 import * as redisCache from './redis';
+import type { PackageCache } from './types';
 
 let cacheProxy: PackageCache;
